+++ server.js
@@ -18,8 +18,15 @@ let seconds = 0;
 let game = new Game();
 let messenger = new Messenger();
 let tree = new Tree();
-let castle = new Building();
 let firstUnit = new Unit();
+
+let castle = new Building();
+castle.define('castle', [
+    {name: 'iron', amount: 32},
+    {name: 'wood', amount: 33},
+    {name: 'clay', amount: 31},
+    {name: 'grain', amount: 34},
+]);
 
 // websockets
 server.on('connection', ws => {
@@ -34,7 +41,11 @@ game.addBuildingTreeAndUnits(tree, [firstUnit]);
 // @todo make this operation atomic
 game.addDemoUser({ username: 'user', password: 'password', });
 game.addVillage('user', new Village('user', 'sensorario\'s village'));
-messenger.setState(game.numberOfVillages())
+
+messenger.setState({
+    numberOfVillages: game.numberOfVillages(),
+    numberOfFields: game.numberOfFields(),
+});
 
 // castle of level 2 unlock units
 game.grandUnitBuildiner({ building: 'castle', level: 2 });